backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
numpy==1.26.4
pandas==2.1.0
google-cloud-bigquery==3.12.0
db-dtypes>=1.1.2
